고객센터 보드들 변경 후 리액트 연결

+++ src/main/java/be/com/bemedicare/board/repository/Board2Repository.java
@@ -5,6 +5,6 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface Board2Repository extends JpaRepository<Board2, Integer> {
+public interface Board2Repository extends JpaRepository<Board2, Long> {
     Page<Board2> findByTitleContaining(String searchKeyword, Pageable pageable);
 }
